pyproject.toml
[tool.poetry]
name = "nonebot_plugin_learning_chat"
version = "0.1.0"
description = "Nonebot2 plugin to learn what your group members say."
authors = ["CMHopeSunshine <277073121@qq.com>"]
keywords = ["nonebot2"]
homepage = "https://github.com/CMHopeSunshine/nonebot-plugin-learning-chat"
readme = "README.md"
license = "AGPL"

[tool.poetry.dependencies]
python = "^3.8"
nonebot2 = "^2.0.0-rc.2"
nonebot-adapter-onebot = "^2.1"
nonebot-plugin-apscheduler = "^0.2.0"
tortoise-orm = "^0.19.2"
jieba = "^0.42.1"
"ruamel.yaml" = "^0.17.21"
amis-python = "^1.0.6"
python-jose = "^3.3.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
